Moved things into 'driver' module, added a couple new pieces of documentation for newly public apis
