+++ nsm-nitro-enclave-utils/src/driver/dev/sign.rs
